Cleaning linear interactive models & started mixed
AIC forward step (the first) does include "religious" because it has to; it starts with all of the predictors by themselves and only adds to it. Does this mean we should only look at the forwad step that starts with the intercept model? Problem sets seem to do the second primarily (?) or start with the backward result, so doesn't necessarily make sense to even include forward step I think unless you think that analysis is helpful. 

I left AIC alone so we can discuss it a bit. I have been trying something with the mixed effects and think I am good to do it and analyze this portion, but take a look / add if you are at a loose end. I'll be finished late tonight with it.